Added tags to recipe and its impl. This makes it easier to access them. Made a helper private in that class. Made a Matcher interface. TagMatcher and IngredientMatcher implement it.
