+++ src/main/java/Matchers/Implementations/IngredientMatcher.java
@@ -1,0 +1,4 @@
+package Matchers.Implementations;
+
+public class IngredientMatcher {
+}
